Keep definitions for web interfaces where needed
