+++ src/CollectingUserTweets.ts
@@ -1,5 +1,14 @@
-import { Tweet } from "./Tweet";
-import { UserTweetsView } from "./UserTweetsView";
+export type Tweet = {
+  author: string;
+  url: string;
+  time: Date;
+  text: string;
+};
+
+export type UserTweetsView = {
+  open: (username: string) => Promise<void>;
+  readNextTweet: () => Promise<Tweet | null>;
+};
 
 export const collectingUserTweets = async (
   tweetsView: UserTweetsView,
