--- src/Tweet.ts
@@ -1,6 +1,0 @@
-export type Tweet = {
-  author: string;
-  url: string;
-  time: Date;
-  text: string;
-};
