--- src/UserProfileView.ts
@@ -1,5 +1,0 @@
-export type UserProfileView = {
-  open: (username: string) => Promise<void>;
-  readName: () => Promise<string>;
-  readBio: () => Promise<string | undefined>;
-};
