--- src/UserTweetsView.ts
@@ -1,6 +1,0 @@
-import { Tweet } from "./Tweet";
-
-export type UserTweetsView = {
-  open: (username: string) => Promise<void>;
-  readNextTweet: () => Promise<Tweet | null>;
-};
